tests/wpt/web-platform-tests/navigation-api/currententrychange-event/history-back-same-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  await navigation.navigate("#foo");
  assert_equals(navigation.entries().length, 2);

  let oncurrententrychange_called = false;
  navigation.oncurrententrychange = t.step_func(e => {
    oncurrententrychange_called = true;
    assert_equals(e.from, navigation.entries()[1]);
    assert_equals(e.navigationType, "traverse");
    assert_equals(navigation.currentEntry.index, 0);
  });
  history.back();
  assert_false(oncurrententrychange_called);
  await new Promise(resolve => window.onpopstate = resolve);
  assert_true(oncurrententrychange_called);
}, "currententrychange fires for history.back()");
</script>
