LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/history-back-same-doc-expected.txt


FAIL currententrychange fires for history.back() assert_equals: expected 3 but got 2
